Remove schuettlebox from functions
After changing the way files are read in the cli from lazy reading to file.read() there was no need to keep sending the schuettelbox to the functions. The schuettelbox is now created for the first time within in the function and is returned in a tuple with the clozed text.

This makes the handling of several files with several functions (adj, nth, etc.) possible.